requirements.txt
flask
pypdf
flask-cors
pyttsx3
reportlab
edge-tts
aiofiles
google-cloud-texttospeech
